Loosen semver ranges
request 2.75.0 depends, via a few different dependency chains, on `hoek`, which is flagged as a moderate security risk by `npm audit`. Widening the semver range for request allows the latest version of hoek - which patches the security flaw - to be picked up. I've also widened the other semver ranges in your dependencies to improve the library's ability to pick up security patches automatically.